moved crossterm to an optional dependency
