+++ Cargo.toml
@@ -11,7 +11,7 @@ members = [
 
 [package]
 name = "ratatui-code-editor"
-version = "0.0.3"
+version = "0.1.0"
 edition = "2024"
 authors = ["vipmax"]
 description = "ratatui-code-editor is A code editor widget for ratatui, built with syntax highlighting powered by Tree-sitter."
@@ -37,7 +37,7 @@ ratatui-code-editor = { path = "." }
 
 [dependencies]
 anyhow = "^1.0"
-crossterm = "~0.29"
+crossterm = { version = "~0.29", optional = true }
 ratatui = "~0.30"
 ropey = "^1.6"
 rust-embed = { version = "^8.7", features = ["include-exclude"] }
@@ -62,3 +62,9 @@ tree-sitter-json = "~0.24.8"
 tree-sitter-toml-ng = "~0.7.0"
 tree-sitter-bash = "~0.25.0"
 tree-sitter-md = "~0.5.1"
+
+[features]
+default = ["crossterm"]
+crossterm = [
+    "dep:crossterm"
+]
